Create taskCard folder and move components for the task card into it
